Respond to feedback on ArrayList
