--- pkg/rancher-components/src/components/Form/ArrayList/index.ts
@@ -1,1 +1,0 @@
-export { default as ArrayList } from './ArrayList.vue';
