--- pkg/rancher-components/src/components/Form/ArrayList/types.d.ts
@@ -1,8 +1,0 @@
-export type ArrayListRow = {
-    value: string;
-}
-
-export type ArrayListData = {
-    lastUpdateWasFromValue: boolean;
-    rows: Array<ArrayListRow>,
-}
