+++ pkg/rancher-components/src/components/Form/index.ts
@@ -3,4 +3,3 @@ export * from './LabeledInput';
 export * from './Radio';
 export * from './TextArea';
 export * from './ToggleSwitch';
-export * from './ArrayList';
