+++ shell/components/form/ArrayListSelect.vue
