+++ shell/edit/persistentvolume/plugins/cephfs.vue
