+++ shell/edit/persistentvolume/plugins/rbd.vue
