+++ stories/components/form/ArrayList.stories.mdx
@@ -61,7 +61,7 @@ The ArrayList is used to input a list of strings. The list is of variable length
 <Source
   language='ts'
   code={`
-     import ArrayList from '@/pkg/rancher-components/src/components/Form/ArrayList/ArrayList';
+     import ArrayList from '@shell/components/form/ArrayList';
   `}
 />
